commit works on push notification, wish list, updated check out, updated search product, registration email notification, checkout email notification,

+++ insantani/app/Http/Controllers/CheckoutController.php
@@ -11,6 +11,10 @@ use Mail;
 use App\User;
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
+use PushNotification;
+use App\WishListModel;
+use App\DeviceTokenModel;
+
 
 class CheckoutController extends Controller
 {
@@ -35,23 +39,65 @@ class CheckoutController extends Controller
                 
 //          
             $failed=array();
+            
+            $failedLocation=array();
             $data=$items['data'];
+           
             for ($i=0; $i<count($data);$i++){
-                $todo=CheckoutModel::create([
-                    'address'=>$data[$i]['address'],
-                    'user_id'=>$data[$i]['user_id'],
-                    'product_id'=>$data[$i]['product_id'],
-                    'productQty'=>$data[$i]['productQty'],
-                    'status'=>'pending'
-                    ]);
+                
                 $todo2=ProductModel::find($data[$i]['product_id']);
+                $todo2->farmer();
+                $data[$i]['product_name']=$todo2->product_name;
+                $data[$i]['farmer_username']=$todo2->farmer_username;
+                $data[$i]['prod_price']=$todo2->prod_price;
+                $farmer=$todo2->farmer;
+                
+                $distance=(6371 * acos(cos(deg2rad($data[$i]['latitude'])) * cos(deg2rad($farmer->latitude)) * cos(deg2rad($farmer->longitude) - deg2rad($data[$i]['longitude'])) + sin(deg2rad($data[$i]['latitude'])) * sin(deg2rad($farmer->latitude))));
+                
+//                $todo2->distance=0;
                 if((($todo2->stock_num)-$data[$i]['productQty'])>=0){
-                    $todo2->stock_num=($todo2->stock_num)-$data[$i]['productQty'];
-                    $todo2->save();
-                    $todo->save();
-                    $todos=ShoppingCartModel::where("user_id","=",$data[$i]['user_id'])
-                                ->where("product_id","=",$data[$i]['product_id']);
-                    $todos->delete();
+                    if ($distance<25){
+                        $todo2->stock_num=($todo2->stock_num)-$data[$i]['productQty'];
+                        $todo2->save();
+                        $todo2->distance=$distance;
+                        if ($todo2->stock_num==0){
+                            
+                            $wishList=WishListModel::where("product_id",'=',$todo2->id)->get();
+        
+                            $deviceList=array();
+                            foreach($wishList as $wish){
+                                $tokens=DeviceTokenModel::where("user_id","=",$wish['user_id'])->get();
+                                foreach($tokens as $token){
+
+                                    array_push($deviceList,PushNotification::Device($token->pluck("device_token")));
+
+                                }
+                            }
+                            $devices= PushNotification::DeviceCollection($deviceList);
+                            $collection=PushNotification::app('appNameAndroid')
+                                            ->to($devices)
+                                            ->send($todo2->product_name.' by '.$todo2->farmer_username.' is currently empty ');
+                             // get response for each device push
+                            foreach ($collection->pushManager as $push) {
+                                $response = $push->getAdapter()->getResponse();
+                            }
+                            
+                        }
+                        $todo=CheckoutModel::create([
+                            'address'=>$data[$i]['address'],
+                            'user_id'=>$data[$i]['user_id'],
+                            'product_id'=>$data[$i]['product_id'],
+                            'productQty'=>$data[$i]['productQty'],
+                            'status'=>'pending'
+                        ]);
+                        $todo->save();
+                        $todos=ShoppingCartModel::where("user_id","=",$data[$i]['user_id'])
+                                                ->where("product_id","=",$data[$i]['product_id']);
+                        $todos->delete();
+                        
+                    }else {
+                        array_push($failedLocation,$data[$i]);
+                    }
                 }else{
                     array_push($failed,$data[$i]);
                 }
@@ -59,12 +105,34 @@ class CheckoutController extends Controller
             
                 
             }
-            if(count($failed)==0){
+//            print_r($items);
+            $items['data']=$data;
+            if(count($failed)==0 && count($failedLocation)==0){
+                
+                $user=User::find($data[0]['user_id']);
+                Mail::send('insantani_order_review', $items, function($message) use($user)
+                        {
+                            $message->to($user->email, $user->email)
+                                    ->subject('Your Order Review');
+                        });
+                
+                
+                
+                
                 return response()->json(['message'=>'success','state'=>'check out'],201);
-            }else{
+            }else if (count($failed)>0 && count($failedLocation)==0){
                 return response()->json(['message'=>'out of stocks',
                                          'state'=>'check out',
-                                         'data'=>$failed],400);
+                                         'data'=>$failed],201);
+            }else if (count($failed)==0 && count($failedLocation)>0){
+                return response()->json(['message'=>'too far from the farmer',
+                                         'state'=>'check out',
+                                         'data'=>$failedLocation],201);
+            }else{
+                return response()->json(['message'=>'too far from the farmer and out of stocks',
+                                         'state'=>'check out',
+                                         'data_location'=>$failedLocation,
+                                         'data_stocks'=>$failed],201);
             }
             
 //
